requirements.txt
python-telegram-bot==21.3
meowbox==1.0.2
python-dotenv
httpx
